Squashed commits: Business name registration fix, Show currently logged in user, Allow comment deletion Display name automatically populated with real name Comment creation / display is fixed

+++ app/controllers/comments_controller.rb
@@ -19,10 +19,12 @@ class CommentsController < ApplicationController
   end
   
   def destroy
-    if @comment.destroy
-      flash[:notice] = "Bookmark removed"
-    else
-      flash[:error] = "An error occurred."
+    if current_user == @comment.user || current_user = @comment.item.user
+      if @comment.destroy
+        flash[:notice] = "Comment removed"
+      else
+        flash[:error] = "An error occurred."
+      end
     end
 
     redirect_to :back
